Implement websocket client
Reminiscent of the owl drawing tutorial [1] this implements the rest of the
bare-bones websocket client implementation. This also removes the dependency on
the autobahn test suite, which will be readded in a later commit - I'm thinking
of splitting it out into a separate integration test as it takes so much
longer.
Nonetheless, the client as of this commit passes all tests (some non-strict) of
the autobahn websocket client test suite (without extensions like compression).

I'm not particularly happy about how big this commit got but the in-progress
commits I had did not lend themselves well to splitting into separate commits
and I don't feel like rebasing to change that.

[1]
1. Draw some lines
2. Draw the rest of the fucking owl