Rename to yawc
This project started as a playground to learn clojure. I wanted to scrape
something so why not learn some clojure while doing so. Rather than actually
doing so I ended up getting sidetracked into implementing a websocket
client to communicate with headless chrome.
The name headless was based on that. As I want to implement the chrome headless
thing in a separate project the name doesn't make sense here anymore. This is a
websocket client. It should be named like one. And yawc is the best I can think
of right now.

TLDR: The name doesn't make sense anymore as this is a websocket client now,
not a chrome headless library.